+++ terraform/util_test.go
@@ -49,56 +49,6 @@ func TestStrSliceContains(t *testing.T) {
 	}
 }
 
-func TestUtilResourceProvider(t *testing.T) {
-	type testCase struct {
-		ResourceName string
-		Alias        string
-		Expected     string
-	}
-
-	tests := []testCase{
-		{
-			// If no alias is provided, the first underscore-separated segment
-			// is assumed to be the provider name.
-			ResourceName: "aws_thing",
-			Alias:        "",
-			Expected:     "aws",
-		},
-		{
-			// If we have more than one underscore then it's the first one that we'll use.
-			ResourceName: "aws_thingy_thing",
-			Alias:        "",
-			Expected:     "aws",
-		},
-		{
-			// A provider can export a resource whose name is just the bare provider name,
-			// e.g. because the provider only has one resource and so any additional
-			// parts would be redundant.
-			ResourceName: "external",
-			Alias:        "",
-			Expected:     "external",
-		},
-		{
-			// Alias always overrides the default extraction of the name
-			ResourceName: "aws_thing",
-			Alias:        "tls.baz",
-			Expected:     "tls.baz",
-		},
-	}
-
-	for _, test := range tests {
-		got := resourceProvider(test.ResourceName, test.Alias)
-		if got != test.Expected {
-			t.Errorf(
-				"(%q, %q) produced %q; want %q",
-				test.ResourceName, test.Alias,
-				got,
-				test.Expected,
-			)
-		}
-	}
-}
-
 func TestUniqueStrings(t *testing.T) {
 	cases := []struct {
 		Input    []string
